+++ src/main/java/net/geforcemods/securitycraft/blocks/FrameBlock.java
@@ -6,9 +6,13 @@ import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.block.AbstractBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.IWaterLoggable;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.fluid.FluidState;
+import net.minecraft.fluid.Fluids;
 import net.minecraft.item.BlockItemUseContext;
 import net.minecraft.item.ItemStack;
+import net.minecraft.state.BooleanProperty;
 import net.minecraft.state.DirectionProperty;
 import net.minecraft.state.StateContainer.Builder;
 import net.minecraft.state.properties.BlockStateProperties;
@@ -25,10 +29,12 @@ import net.minecraft.util.math.shapes.VoxelShape;
 import net.minecraft.util.math.shapes.VoxelShapes;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.IBlockReader;
+import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
 
-public class FrameBlock extends OwnableBlock {
+public class FrameBlock extends OwnableBlock implements IWaterLoggable {
 	public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
+	public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
 	private static final VoxelShape SHAPE_NORTH = VoxelShapes.joinUnoptimized(VoxelShapes.block(), Block.box(2, 2, 0, 14, 14, 1), IBooleanFunction.ONLY_FIRST);
 	private static final VoxelShape SHAPE_EAST = VoxelShapes.joinUnoptimized(VoxelShapes.block(), Block.box(15, 2, 2, 16, 14, 14), IBooleanFunction.ONLY_FIRST);
 	private static final VoxelShape SHAPE_SOUTH = VoxelShapes.joinUnoptimized(VoxelShapes.block(), Block.box(2, 2, 15, 14, 14, 16), IBooleanFunction.ONLY_FIRST);
@@ -36,7 +42,7 @@ public class FrameBlock extends OwnableBlock {
 
 	public FrameBlock(AbstractBlock.Properties properties) {
 		super(properties);
-		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH));
+		registerDefaultState(stateDefinition.any().setValue(FACING, Direction.NORTH).setValue(WATERLOGGED, false));
 	}
 
 	@Override
@@ -68,13 +74,26 @@ public class FrameBlock extends OwnableBlock {
 	}
 
 	@Override
+	public BlockState updateShape(BlockState state, Direction facing, BlockState facingState, IWorld level, BlockPos currentPos, BlockPos facingPos) {
+		if (state.getValue(WATERLOGGED))
+			level.getLiquidTicks().scheduleTick(currentPos, Fluids.WATER, Fluids.WATER.getTickDelay(level));
+
+		return super.updateShape(state, facing, facingState, level, currentPos, facingPos);
+	}
+
+	@Override
+	public FluidState getFluidState(BlockState state) {
+		return state.getValue(WATERLOGGED) ? Fluids.WATER.getSource(false) : super.getFluidState(state);
+	}
+
+	@Override
 	public BlockState getStateForPlacement(BlockItemUseContext ctx) {
-		return defaultBlockState().setValue(FACING, ctx.getPlayer().getDirection().getOpposite());
+		return defaultBlockState().setValue(FACING, ctx.getPlayer().getDirection().getOpposite()).setValue(WATERLOGGED, ctx.getLevel().getFluidState(ctx.getClickedPos()).getType() == Fluids.WATER);
 	}
 
 	@Override
 	protected void createBlockStateDefinition(Builder<Block, BlockState> builder) {
-		builder.add(FACING);
+		builder.add(FACING, WATERLOGGED);
 	}
 
 	@Override